+++ src/api/pbx.ts
@@ -13,6 +13,7 @@ import { bundleIdentifier, fcmApplicationId } from '../config'
 import { embedApi } from '../embed/embedApi'
 import type { Account } from '../stores/accountStore'
 import { accountStore } from '../stores/accountStore'
+import { authPBX } from '../stores/AuthPBX'
 import { getAuthStore, waitPbx } from '../stores/authStore'
 import type { PbxUser, Phonebook } from '../stores/contactStore'
 import { intl } from '../stores/intl'
@@ -61,6 +62,7 @@ export class PBX extends EventEmitter {
       tenant: a.pbxTenant,
       login_user: a.pbxUsername,
       login_password: a.pbxPassword,
+      phone_idx: a.pbxPhoneIndex,
       _wn: d.accessToken,
       park: a.parks || [],
       voicemail: 'self',
@@ -196,6 +198,7 @@ export class PBX extends EventEmitter {
     setListenerWithEmbed('notify_callrecording', this.onCallRecording)
     setListenerWithEmbed('notify_voicemail', this.onVoicemail)
     setListenerWithEmbed('notify_status', this.onUserStatus)
+    setListenerWithEmbed('notify_pal', this.onUserLoginOtherDevices)
     // fire pending events
     pendingClose.forEach(this.onClose)
     pendingError.forEach(this.onError)
@@ -283,6 +286,15 @@ export class PBX extends EventEmitter {
     }
   }
 
+  private onUserLoginOtherDevices = (e: PbxEvent['pal']) => {
+    if (!e) {
+      return
+    }
+    if (e.code === 1 && e.message === 'ANOTHER_LOGIN') {
+      getAuthStore().pbxLoginFromAnotherPlace = true
+      authPBX.dispose()
+    }
+  }
   disconnect = () => {
     if (this.client) {
       this.client.close()
@@ -321,6 +333,7 @@ export class PBX extends EventEmitter {
 
     const as = getAuthStore()
     as.pbxConfig = config
+    as.setRecentCallsMax(config['webphone.recents.max'])
 
     // the custom page only load at the first time the tab is shown after you log in
     //    even after re-connected it, don't refresh it again
@@ -807,7 +820,7 @@ export class PBX extends EventEmitter {
 
 export const pbx = new PBX()
 // ----------------------------------------------------------------------------
-// Parse resource line data
+// parse resource line data
 const _parseResourceLines = (l: string | undefined) => {
   const as = getAuthStore()
   if (!l) {