+++ src/components/App.tsx
@@ -264,8 +264,9 @@ export const App = observer(() => {
     sipConnectingOrFailure,
     ucConnectingOrFailure,
     ucLoginFromAnotherPlace,
+    pbxLoginFromAnotherPlace,
     signedInId,
-    resetFailureStateIncludeUcLoginFromAnotherPlace,
+    resetFailureStateIncludePbxOrUc,
   } = getAuthStore()
 
   const serviceConnectingOrFailure = pbxConnectingOrFailure()
@@ -276,15 +277,20 @@ export const App = observer(() => {
         ? 'UC'
         : ''
   const isFailure = isConnFailure()
-
   const connMessage =
-    isFailure && ucLoginFromAnotherPlace
-      ? intl`UC signed in from another location`
-      : !serviceConnectingOrFailure
-        ? ''
-        : isFailure
-          ? intl`${serviceConnectingOrFailure} connection failed`
-          : intl`Connecting to ${serviceConnectingOrFailure}...`
+    isFailure && pbxLoginFromAnotherPlace
+      ? intl`Logged in from another location as the same phone`
+      : isFailure && ucLoginFromAnotherPlace
+        ? intl`UC signed in from another location`
+        : !serviceConnectingOrFailure
+          ? ''
+          : isFailure
+            ? intl`${serviceConnectingOrFailure} connection failed`
+            : intl`Connecting to ${serviceConnectingOrFailure}...`
+
+  const onPressConnMessage = isFailure
+    ? resetFailureStateIncludePbxOrUc
+    : undefined
 
   const cp = getAuthStore().listCustomPage[0]
 
@@ -301,11 +307,7 @@ export const App = observer(() => {
         >
           <RnTouchableOpacity
             style={css.App_ConnectionStatusInner}
-            onPress={
-              isFailure
-                ? resetFailureStateIncludeUcLoginFromAnotherPlace
-                : undefined
-            }
+            onPress={onPressConnMessage}
           >
             <RnText small white>
               {connMessage}
@@ -331,7 +333,7 @@ export const App = observer(() => {
         {isFailure && (
           <RnTouchableOpacity
             style={css.App_ConnectionStatusIncreaseTouchSize}
-            onPress={resetFailureStateIncludeUcLoginFromAnotherPlace}
+            onPress={onPressConnMessage}
           />
         )}
       </View>