+++ src/pages/PageCallRecents.tsx
@@ -2,19 +2,25 @@ import { observer } from 'mobx-react'
 import moment from 'moment'
 import { Component } from 'react'
 import type { NativeEventSubscription } from 'react-native'
-import { AppState, Platform } from 'react-native'
+import { AppState, Platform, StyleSheet } from 'react-native'
 
 import { mdiMagnify, mdiPhone, mdiVideo } from '../assets/icons'
 import { UserItem } from '../components/ContactUserItem'
 import { Field } from '../components/Field'
 import { Layout } from '../components/Layout'
-import type { RecentCall } from '../stores/accountStore'
-import { accountStore } from '../stores/accountStore'
+import { RnText } from '../components/RnText'
+import { RnTouchableOpacity } from '../components/RnTouchableOpacity'
 import { getAuthStore } from '../stores/authStore'
 import { getCallStore } from '../stores/callStore'
 import { contactStore } from '../stores/contactStore'
 import { intl } from '../stores/intl'
 import { PushNotification } from '../utils/PushNotification'
+
+const css = StyleSheet.create({
+  Loading: {
+    marginTop: 20,
+  },
+})
 
 @observer
 export class PageCallRecents extends Component {
@@ -30,17 +36,14 @@ export class PageCallRecents extends Component {
       h()
       this.appStateSubscription = AppState.addEventListener('change', h)
     }
+    getAuthStore().rcFirstTimeLoadData()
   }
   componentWillUnmount = () => {
     this.appStateSubscription?.remove()
+    getAuthStore().cRecentCalls = []
+    getAuthStore().rcPage = 0
   }
 
-  isMatchUser = (call: RecentCall) => {
-    if (call.partyNumber.includes(contactStore.callSearchRecents)) {
-      return call.id
-    }
-    return ''
-  }
   getAvatar = (id: string) => {
     const ucUser = contactStore.getUcUserById(id) || {}
     return {
@@ -48,38 +51,13 @@ export class PageCallRecents extends Component {
       avatar: ucUser.avatar,
     }
   }
-  getMatchedCalls = () => {
-    const as = getAuthStore()
-    const ca = as.getCurrentAccount()
-    const d = as.getCurrentData()
-    if (!d && ca) {
-      // trigger async update
-      accountStore.findDataWithDefault(ca)
-    }
-    const calls = d?.recentCalls.filter(this.isMatchUser) || []
-
-    // backward compatibility to remove invalid items from the previous versions
-    const filteredCalls = calls.filter(
-      c =>
-        typeof c.created === 'string' &&
-        // format: HH:mm - MMM D
-        ((c.created + '').length === 13 || (c.created + '').length === 14),
-    )
-    const today = moment().format('MMM D')
-    return filteredCalls.map(c => ({
-      ...c,
-      created:
-        (c.created + '').replace(` - ${today}`, '') +
-        `${as.resourceLines.length && c.lineValue ? '  ⭞ ' + `${c?.lineLabel ? c.lineLabel + ':' : ''} ${c.lineValue}` : ''}`,
-    }))
-  }
 
   render() {
-    const calls = this.getMatchedCalls()
-    const auth = getAuthStore()
+    const as = getAuthStore()
+    const calls = as.cRecentCalls
     return (
       <Layout
-        description={intl`Voicemail and recent calls`}
+        description={intl`Recent calls`}
         menu='call'
         subMenu='recents'
         title={intl`Recents`}
@@ -90,33 +68,51 @@ export class PageCallRecents extends Component {
           onValueChange={(v: string) => {
             // TODO use debounced value to perform data filter
             contactStore.callSearchRecents = v
+            as.rcSearchRecentCall()
           }}
           value={contactStore.callSearchRecents}
         />
-        <Field isGroup label={intl`RECENT CALLS (${calls.length})`} />
-        {calls.map((c, i) => (
-          <UserItem
-            iconFuncs={[
-              () => getCallStore().startVideoCall(c.partyNumber),
-              () =>
-                getCallStore().startCall(
-                  c.partyNumber,
-                  (auth.resourceLines.length &&
-                    c.lineValue && {
-                      extraHeaders: [`X-PBX-RPI: ${c.lineValue}`],
-                    }) ||
-                    undefined,
-                ),
-            ]}
-            {...contactStore.getUcUserById(c.partyNumber)}
-            icons={[mdiVideo, mdiPhone]}
-            isRecentCall
-            canChat={getAuthStore().getCurrentAccount()?.ucEnabled}
-            key={i}
-            {...this.getAvatar(c.partyNumber)}
-            {...c}
-          />
-        ))}
+        <Field isGroup label={intl`RECENT CALLS (${as.rcCount})`} />
+        {calls.map((c, i) => {
+          const today = moment().format('MMM D')
+          Object.assign(c, {
+            created: (c.created + '').replace(` - ${today}`, ''),
+          })
+          return (
+            <UserItem
+              iconFuncs={[
+                () => getCallStore().startVideoCall(c.partyNumber),
+                () => getCallStore().startCall(c.partyNumber),
+              ]}
+              {...contactStore.getUcUserById(c.partyNumber)}
+              icons={[mdiVideo, mdiPhone]}
+              isRecentCall
+              canTouch={as.getCurrentAccount()?.ucEnabled}
+              key={i}
+              {...this.getAvatar(c.partyNumber)}
+              {...c}
+            />
+          )
+        })}
+        {as.rcLoading ? (
+          <RnText
+            style={css.Loading}
+            warning
+            small
+            normal
+            center
+          >{intl`Loading...`}</RnText>
+        ) : as.cRecentCalls.length < as.rcCount ? (
+          <RnTouchableOpacity onPress={as.rcLoadMore}>
+            <RnText
+              style={css.Loading}
+              primary
+              small
+              normal
+              center
+            >{intl`Load more recent calls`}</RnText>
+          </RnTouchableOpacity>
+        ) : null}
       </Layout>
     )
   }