+++ src/stores/callStore2.ts
@@ -320,11 +320,15 @@ export class CallStore {
       }
       return
     }
+
     //
     // construct a new call
     const c = new Call(this)
     Object.assign(c, p)
-
+    // clear start call interval timer when the outgoing call is created
+    if (Platform.OS === 'web' && !c.incoming) {
+      this.clearStartCallIntervalTimer()
+    }
     // get Avatar and Username of phoneappli
     const ca = auth.getCurrentAccount()
     if (auth.phoneappliEnabled() && !c.incoming && ca) {
@@ -479,7 +483,7 @@ export class CallStore {
     const index = extraHeaders.findIndex(header =>
       header.startsWith('X-PBX-RPI:'),
     )
-    // If it allows calling without a value ( no-line ), then make a call without a line resource.
+    // if it allows calling without a value `no-line`, then make a call without a line resource.
     if (resourceLines[0].key === 'no-line' && index !== -1) {
       extraHeaders.splice(index, 1)
       return extraHeaders
@@ -534,7 +538,7 @@ export class CallStore {
     })
   }
   startCall: MakeCallFn = async (number: string, ...args) => {
-    // Make sure sip is ready before make call
+    // make sure sip is ready before make call
     if (getAuthStore().sipState !== 'success') {
       return
     }